*) Changes from Dre - Removed: lot of useless casts (mainly related to GameNPC.eFlags - Changed: removed useless t/c and added new ones
*) Changes from me:
- Changed: the trainer window call is now in GameTrainer class
- Removed: all class call to ttainer window
- Changed: missingg experience and freelevel 'll only be given by your class trainer or an universal trainer.

Theses changes are aiming the 1.105 patch. For your universal trainer, using eCharacter.Unknown will always call the trainer window, while others classes trainers will only work with corresponding player class.

+++ GameServer/ai/brain/StandardMobBrain.cs
@@ -779,7 +779,7 @@ namespace DOL.AI.Brain
 					// first check to see if the healer is in our aggrolist so we don't go attacking anyone who heals
 					if (m_aggroTable[(GameLiving)eArgs.HealSource] != null)
 					{
-						if (eArgs.HealSource is GamePlayer || (eArgs.HealSource is GameNPC && (((GameNPC)eArgs.HealSource).Flags & (uint)GameNPC.eFlags.PEACE) == 0))
+						if (eArgs.HealSource is GamePlayer || (eArgs.HealSource is GameNPC && (((GameNPC)eArgs.HealSource).Flags & GameNPC.eFlags.PEACE) == 0))
 						{
 							AddToAggroList((GameLiving)eArgs.HealSource, eArgs.HealAmount);
 						}
