+++ GameServer/serverrules/NormalServerRules.cs
@@ -123,11 +123,11 @@ namespace DOL.GS.ServerRules
 
 			//Peace flag NPCs are same realm
 			if (target is GameNPC)
-				if ((((GameNPC)target).Flags & (uint)GameNPC.eFlags.PEACE) != 0)
+				if ((((GameNPC)target).Flags & GameNPC.eFlags.PEACE) != 0)
 					return true;
 
 			if (source is GameNPC)
-				if ((((GameNPC)source).Flags & (uint)GameNPC.eFlags.PEACE) != 0)
+				if ((((GameNPC)source).Flags & GameNPC.eFlags.PEACE) != 0)
 					return true;
 
 			if(source.Realm != target.Realm)
@@ -186,11 +186,11 @@ namespace DOL.GS.ServerRules
 
 			//Peace flag NPCs can trade with everyone
 			if (target is GameNPC)
-				if ((((GameNPC)target).Flags & (uint)GameNPC.eFlags.PEACE) != 0)
+				if ((((GameNPC)target).Flags & GameNPC.eFlags.PEACE) != 0)
 					return true;
 
 			if (source is GameNPC)
-				if ((((GameNPC)source).Flags & (uint)GameNPC.eFlags.PEACE) != 0)
+				if ((((GameNPC)source).Flags & GameNPC.eFlags.PEACE) != 0)
 					return true;
 
 			if(source.Realm != target.Realm)
@@ -212,7 +212,7 @@ namespace DOL.GS.ServerRules
 			//Peace flag NPCs can be understood by everyone
 
 			if (source is GameNPC)
-				if ((((GameNPC)source).Flags & (uint)GameNPC.eFlags.PEACE) != 0)
+				if ((((GameNPC)source).Flags & GameNPC.eFlags.PEACE) != 0)
 					return true;
 
 			if(source.Realm > 0 && source.Realm != target.Realm) return false;
